requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
maprule==1.1.0
metatron==0.5
mypy==0.931
mypy-extensions==0.4.3
Pillow==9.0.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
requests==2.27.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tomli==2.0.1
Twisted==22.1.0
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.8
whitenoise==6.0.0
zope.interface==5.4.0
